% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btf_master.R
\docType{data}
\name{btf_master}
\alias{btf_master}
\title{Bethel Test Fishery CPUE Records}
\format{
See \code{vignette("datasets", package = "KuskoHarvData")}.
}
\source{
This CPUE index is used as an index of in-river abundance, species composition, and -- at the end of the season -- run timing.
It can be found on the \href{https://www.adfg.alaska.gov/index.cfm?adfg=commercialbyareakuskokwim.btf}{ADF&G Bethel Test Fishery website}.
Sampling methods are described in Bue and Lipka (\href{http://www.adfg.alaska.gov/FedAidPDFs/FDS16-05.pdf}{2016}).
}
\usage{
btf_master
}
\description{
Data set containing the daily catch-per-unit-effort index generated by the Bethel Test Fishery,
operated by the Alaska Department of Fish and Game.
}
\note{
This data set needs to be updated manually,
see Step 5 in \code{vignette("updating-data", package = "KuskoHarvData")}.
}
\keyword{datasets}
