% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_est_all.R
\docType{data}
\name{effort_est_all}
\alias{effort_est_all}
\title{Compiled Effort Estimates}
\format{
See \code{vignette("datasets", package = "KuskoHarvData")}.
}
\source{
The number of trips is estimated by combining aerial boat counts
(found in \verb{data("flight_data_all", package = "KuskoHarvData"}) and
access point completed trip interviews (found in \code{data("interview_data_all")}).
The estimator is described in Staton et al. (In Review).
}
\usage{
effort_est_all
}
\description{
Data set containing all compiled estimates of the number of drift gillnet trips
that occurred in each drift gillnet subsistence fishery opener, as estimated by
the Lower Kuskokwim River In-season Subsistence Salmon Harvest Monitoring Program.
}
\note{
This data set will be updated automatically when adding new interview and flight data,
see Steps 3 and 6 in \code{vignette("updating-data", package = "KuskoHarvData")}.
}
\keyword{datasets}
