% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_estimate_master.R
\docType{data}
\name{effort_estimate_master}
\alias{effort_estimate_master}
\title{Effort Estimates}
\format{
A data frame with rows representing a unique date- and area-specific effort estimate.
}
\usage{
effort_estimate_master
}
\description{
A data set containing all compiled estimates of the number of
drift net trips that occurred in each subsistence harvest opportunity.
}
\keyword{datasets}
