% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_data_master.R
\docType{data}
\name{flight_data_master}
\alias{flight_data_master}
\title{Aerial Count Records}
\format{
A data frame with rows representing individual flights and columns representing variables like the start/end times of the flight and the area-specific counts.
}
\usage{
flight_data_master
}
\description{
A data set containing all aerial counts of drift net boats collected by the Kuskokwim River
in-season subsistence salmon harvest monitoring program.
}
\keyword{datasets}
