% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harv_est_all.R
\docType{data}
\name{harv_est_all}
\alias{harv_est_all}
\title{Compiled Harvest Estimates}
\format{
A data frame with rows representing a unique date/species/area/summary statistic combination.
}
\usage{
harv_est_all
}
\description{
Data set containing all compiled estimates of harvest by drift gillnet boats generated
by the Lower Kuskokwim River In-season Subsistence Salmon Harvest Monitoring Program.
}
\keyword{datasets}
