% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_estimate_master.R
\docType{data}
\name{harvest_estimate_master}
\alias{harvest_estimate_master}
\title{Compiled Harvest Estimates}
\format{
A data frame with rows representing a unique date/species/area/summary statistic combination.
}
\usage{
harvest_estimate_master
}
\description{
Data set containing all compiled estimates of harvest by drift gillnet boats generated
by the Lower Kuskokwim River In-season Subsistence Salmon Harvest Monitoring Program.
}
\keyword{datasets}
