% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interview_data_master.R
\docType{data}
\name{interview_data_master}
\alias{interview_data_master}
\title{Interview Records}
\format{
A data frame with rows representing individual interviews as rows and question responses as columns.
Also includes a set of logical flags that denote which records are suitable for which purposes from an estimation standpoint.
}
\usage{
interview_data_master
}
\description{
A data set containing all drift net interview data collected by the Kuskokwim River
in-season subsistence salmon harvest monitoring program.
}
\keyword{datasets}
