% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{meta}
\alias{meta}
\title{Meta-data for each opener with monitoring data}
\format{
A data frame with rows representing a unique day in which the fishery was open and monitoring occurred.
Variables include:
\itemize{
\item \code{start}: the earliest time fishing was allowed that day (stored as a datetime class)
\item \code{end}: the latest time fishing was allowed that day (stored as a datetime class)
\item \code{flights_planned}: the number of flights planned to be flown that day
\item \code{flights_flown}: the number of flights actually flown that day
\item \code{announcement}: the identifier of the official document announcing that fishing would be allowed that day
}
}
\usage{
meta
}
\description{
A data set containing basic information about each opener with interview and/or flight data
}
\keyword{datasets}
