% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openers_all.R
\docType{data}
\name{openers_all}
\alias{openers_all}
\title{Opener Metadata}
\format{
See \code{vignette("datasets", package = "KuskoHarvData")}.
}
\usage{
openers_all
}
\description{
Data set containing basic information about each opener with interview and/or flight data
}
\note{
This data set must be updated manually,
see Step 4 in \code{vignette("updating-data", package = "KuskoHarvData")}.
}
\keyword{datasets}
