% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PABE_data_all.R
\docType{data}
\name{PABE_data_all}
\alias{PABE_data_all}
\title{Daily PABE Weather Summaries}
\format{
See \code{vignette("datasets", package = "KuskoHarvData")}.
}
\source{
This data set relies on the R package \href{https://docs.ropensci.org/riem/index.html}{'riem'}' to query the ASOS
database (airports), hosted by the \href{https://mesonet.agron.iastate.edu/request/download.phtml?network=IN__ASOS}{Iowa State University Environmental Mesonet}.
The weather station used was the Bethel Airport (\code{"PABE"}).
}
\usage{
PABE_data_all
}
\description{
Data set containing daily weather data summaries collected at the Bethel
airport.
}
\note{
This data set will be updated automatically when compiling updating the exported data objects for the package,
see Step 6 in \code{vignette("updating-data", package = "KuskoHarvData")}.
}
\keyword{datasets}
