% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-predictive-data.R
\name{prepare_regression_data}
\alias{prepare_regression_data}
\title{Prepare a data set containing predictor and response variables for regression modeling}
\usage{
prepare_regression_data(dates = NULL, na.omit = TRUE, ...)
}
\arguments{
\item{dates}{Date object}

\item{na.omit}{Logical; if \code{TRUE} (the default), any rows with an \code{NA} value for any variable will be discarded.}

\item{...}{Optional arguments passed to \code{\link[=prepare_response_vars]{prepare_response_vars()}}}
}
\value{
A data frame with rows for individual dates and columns for several variables to be used in regression modeling.
See \code{\link[=prepare_predictor_vars]{prepare_predictor_vars()}} and \code{\link[=prepare_response_vars]{prepare_response_vars()}} for variable definitions.
In addition to those variables, there is also a \code{period} variable (see \code{\link[=get_period]{get_period()}}).
}
\description{
A wrapper around \code{\link[=prepare_predictor_vars]{prepare_predictor_vars()}} and \code{\link[=prepare_response_vars]{prepare_response_vars()}} for one-line
regression data preparation.
}
