% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_data_all.R
\docType{data}
\name{sonar_data_all}
\alias{sonar_data_all}
\title{Kuskokwim River Sonar Count Records}
\format{
See \code{vignette("datasets", package = "KuskoHarvData")}.
}
\source{
This data set is used as an index of in-river abundance, species composition, and -- at the end of the season -- run timing.
It can be found on the \href{https://www.adfg.alaska.gov/index.cfm?adfg=commercialbyareakuskokwim.emihd}{ADF&G Kuskokwim Escapement Fish Counts website}.
Sampling methods are described in Birchfield and Fletcher (\href{https://www.adfg.alaska.gov/FedAidPDFs/FDS25-31.pdf}{2025}).
}
\usage{
sonar_data_all
}
\description{
Data set containing the daily counts generated by the Kuskokwim River sonar,
operated by the Alaska Department of Fish and Game.
}
\note{
This data set needs to be updated manually,
see Step 5 in \code{vignette("updating-data", package = "KuskoHarvData")}.
}
\keyword{datasets}
