% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-btf.R
\name{summarize_btf}
\alias{summarize_btf}
\title{Summarize Bethel Test Fishery CPUE data}
\usage{
summarize_btf(date, stat, plus_minus = 0)
}
\arguments{
\item{date}{Datetime; the date to summarize. Must be of length 1.
If \code{plus_minus > 0}, becomes the center of window across which the average is calculated.}

\item{stat}{Character; the summary stat/species combination to return. Must be of length 1.
Accepted values are:
\itemize{
\item \code{"chinook_cpue"} -- returns the daily CPUE of Chinook salmon, averaged over a window if \code{plus_minus > 0}
\item \code{"chum_cpue"}
\item \code{"sockeye_cpue"}
\item \code{"total_cpue"}
\item \code{"chinook_comp"} -- returns proportion of all daily CPUE made up of Chinook salmon, averaged over a window if \code{plus_minus > 0}
\item \code{"chum_comp"}
\item \code{"sockeye_comp"}
}}

\item{plus_minus}{Numeric; the number of days before and after \code{date} to include in the average statistic returned. Must be of length 1.
\code{plus_minus = 0} (default) will return the observed daily value, \code{plus_minus = 1} will return a three day average with \code{date} as the center day.
Greater values of \code{plus_minus} will introduce more smoothing of between-day variability.}
}
\description{
Obtains a species-specific summary in terms of either proportional composition or daily CPUE
with the ability to average over a several day window to smooth out daily variability.
}
