% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-sonar.R
\name{summarize_sonar}
\alias{summarize_sonar}
\title{Summarize Sonar Count data}
\usage{
summarize_sonar(date, stat, plus_minus = 0)
}
\arguments{
\item{date}{Datetime; the date to summarize. Must be of length 1.
If \code{plus_minus > 0}, becomes the center of window across which the average is calculated.}

\item{stat}{Character; the summary stat/species combination to return. Must be of length 1.
Accepted values are:
\itemize{
\item \code{"chinook_count"} -- returns the daily count of Chinook salmon, averaged over a window if \code{plus_minus > 0}
\item \code{"chum_count"}
\item \code{"sockeye_count"}
\item \code{"total_count"}
\item \code{"chinook_comp"} -- returns proportion of all daily counts made up of Chinook salmon, averaged over a window if \code{plus_minus > 0}
\item \code{"chum_comp"}
\item \code{"sockeye_comp"}
}}

\item{plus_minus}{Numeric; the number of days before and after \code{date} to include in the average statistic returned. Must be of length 1.
\code{plus_minus = 0} (default) will return the observed daily value, \code{plus_minus = 1} will return a three day average with \code{date} as the center day.
Greater values of \code{plus_minus} will introduce more smoothing of between-day variability.}
}
\description{
Obtains a species-specific summary in terms of either proportional composition or daily count
with the ability to average over a several day window to smooth out daily variability.
}
\note{
If the requested date is not found in the sonar data (e.g., 2016, 2017, late-season in 2024) an \code{NA} value will be returned.
}
