% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btf_master.R
\docType{data}
\name{btf_master}
\alias{btf_master}
\title{Bethel Test Fishery CPUE Data}
\format{
A data frame with rows representing a unique day- and species-specific record.
Variables include:
\itemize{
\item \code{date}: the date sampling occurred; only dates between June 1 and August 24 in years after and including 2016 are included
\item \code{species}: the species for which \code{cpue} and \code{ccpue} apply; only Chinook, chum, and sockeye salmon are included
\item \code{cpue}: the daily value of catch-per-unit-effort
\item \code{ccpue}: the cumulative value of catch-per-unit effort
}
}
\usage{
btf_master
}
\description{
A data set containing the daily catch-per-unit-effort index generated by the Bethel Test Fishery,
operated by the Alaska Department of Fish and Game
}
\note{
This data set will need to be updated regularly (at least annually).
It can be found on (this ADF&G website)\link{https://www.adfg.alaska.gov/index.cfm?adfg=commercialbyareakuskokwim.btf}.
}
\keyword{datasets}
