% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_estimate_master.R
\docType{data}
\name{harvest_estimate_master}
\alias{harvest_estimate_master}
\title{Harvest Estimates}
\format{
A data frame with rows representing a unique date/species/area/summary statistic combination.
}
\usage{
harvest_estimate_master
}
\description{
A data set containing all compiled estimates of harvest by drift nets generated
by the Kuskokwim River in-season subsistence salmon harvest monitoring program.
}
\keyword{datasets}
