% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-predictive-data.R
\name{prepare_predictor_vars}
\alias{prepare_predictor_vars}
\title{Prepare a data set containing predictor variables for regression modeling}
\usage{
prepare_predictor_vars(dates = NULL)
}
\arguments{
\item{dates}{Datetime; a vector of historical fishing dates to include.
If \code{NULL} (the default) all dates included in the meta data set will be included.}
}
\value{
A data frame with rows for individual dates and columns for:
\itemize{
\item \code{date}: the date of the fishing day
\item \code{year}: the year of the fishing day
\item \code{day}: days past May 31st in the year of fishing
\item \code{hours_open}: the number of hours that day fishing was allowed that day
\item \code{not_first_day}: \code{TRUE} if the day is not the first day of fishing in a consecutive set of fishing days
\item \code{weekend}: \code{TRUE} if the fishing day occurred on Saturday or Sunday
\item \code{p_before_noon}: fraction of the allowed fishing hours that occurred before noon that day
\item \code{total_btf_cpue}: daily catch-per-unit-effort from the Bethel Test Fishery, averaged over a three day period where \code{date} is the second day
\item \code{chinook_btf_comp}: daily proportional composition of Chinook salmon from the Bethel Test Fishery, averaged over a three day period where \code{date} is the second day
\item \code{chum_btf_comp}: daily proportional composition of chum salmon from the Bethel Test Fishery, averaged over a three day period where \code{date} is the second day
\item \code{sockeye_btf_comp}: daily proportional composition of sockeye salmon from the Bethel Test Fishery, averaged over a three day period where \code{date} is the second day
\item \code{mean_temp}: daily average air temperature, in degrees Fahrenheit
\item \code{mean_relh}: daily average percent relative humidity
\item \code{precip}: total daily precipitation, in inches
}
}
\description{
Quickly summarizes various date/time and Bethel Test Fishery summaries
for fitting in regression models.
}
