% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-predictive-data.R
\name{prepare_response_vars}
\alias{prepare_response_vars}
\title{Prepare a data set containing response variables for regression modeling}
\usage{
prepare_response_vars(
  dates = NULL,
  comp_species = c("chinook", "chum", "sockeye"),
  cpt_species = "total",
  include_transformed = TRUE
)
}
\arguments{
\item{dates}{Datetime; a vector of historical fishing dates to include.
If \code{NULL} (the default) all dates included in the meta data set will be included.}

\item{comp_species}{Character; a vector specifying the species to calculate proportional contribution for.
Accepted options are any combinations of \code{"chinook"}, \code{"chum"}, and \code{"sockeye"}}

\item{cpt_species}{Character; a vector specifying the species to calculate catch rate (catch/trip) for.
Accepted options are any combinations of \code{"chinook"}, \code{"chum"}, \code{"sockeye"}, and \code{"total"}.}

\item{include_transformed}{Logical; if \code{TRUE} (the default), additional columns will be returned on a transformed scale.
Effort and catch rate variables will be log-transformed and composition variables will be logit-transformed.}
}
\value{
A data frame with rows for individual dates and columns for date, effort, and composition/catch rate for the requested species.
All values are calculated for drift nets only, using the mean harvest estimates, and all geographic strata.
}
\description{
Extracts total effort, species composition, and catch rate data from historical
fishing days for fitting in regression models.
}
